+++ R/extend.R
@@ -340,6 +340,7 @@ columns_used.relop_extend <- function (x, ...,
 to_sql.relop_extend <- function (x,
                                  db,
                                  ...,
+                                 limit = NULL,
                                  source_limit = NULL,
                                  indent_level = 0,
                                  tnum = mk_tmp_name_source('tsql'),
@@ -354,8 +355,13 @@ to_sql.relop_extend <- function (x,
   # work on query
   using <- calc_used_relop_extend(x,
                                   using = using)
+  qlimit = limit
+  if(!getDBOption(db, "use_pass_limit", TRUE)) {
+    qlimit = NULL
+  }
   subsql_list <- to_sql(x$source[[1]],
                         db = db,
+                        limit = qlimit,
                         source_limit = source_limit,
                         indent_level = indent_level + 1,
                         tnum = tnum,
@@ -422,6 +428,10 @@ to_sql.relop_extend <- function (x,
               prefix, "FROM (\n",
               subsql, "\n",
               prefix, " ) ", tab)
+  if(!is.null(limit)) {
+    q <- paste(q, "LIMIT",
+               format(ceiling(limit), scientific = FALSE))
+  }
   if(append_cr) {
     q <- paste0(q, "\n")
   }