+++ R/non_sql_node.R
@@ -141,6 +141,7 @@ columns_used.relop_non_sql <- function (x, ...,
 to_sql.relop_non_sql <- function (x,
                                   db,
                                   ...,
+                                  limit = NULL,
                                   source_limit = NULL,
                                   indent_level = 0,
                                   tnum = mk_tmp_name_source('tsql'),
@@ -155,12 +156,10 @@ to_sql.relop_non_sql <- function (x,
                     using = NULL)
   nsubsql <- length(subsql)
   # non-SQL nodes must always be surrounded by SQL on both sides
-  step1 <- list(paste0("CREATE ",
-                       ifelse(x$termporary, "TERMPORARY", ""),
-                       " TABLE ",
-                       quote_identifier(db, x$incoming_table_name),
-                       " AS ",
-                       subsql[[nsubsql]]))
+  step1 <- materialize_sql_statement(db,
+                                     subsql[[nsubsql]],
+                                     x$incoming_table_name,
+                                     temporary = x$temporary)
   nsql_step <- list(display_form = x$display_form,
                     incoming_table_name = x$incoming_table_name,
                     outgoing_table_name = x$outgoing_table_name,
@@ -168,8 +167,13 @@ to_sql.relop_non_sql <- function (x,
                     f = x$f)
   class(nsql_step) <- "rquery_non_sql_step"
   step2 <- list(nsql_step)
+  qlimit = limit
+  if(!getDBOption(db, "use_pass_limit", TRUE)) {
+    qlimit = NULL
+  }
   step3 <- list(to_sql(table_source(x$outgoing_table_name, column_names(x)),
                        db = db,
+                       limit = qlimit,
                        source_limit = source_limit,
                        indent_level = indent_level + 1,
                        tnum = tnum,