+++ R/select_columns.R
@@ -106,6 +106,7 @@ columns_used.relop_select_columns <- function (x, ...,
 to_sql.relop_select_columns <- function (x,
                                          db,
                                          ...,
+                                         limit = NULL,
                                          source_limit = NULL,
                                          indent_level = 0,
                                          tnum = mk_tmp_name_source('tsql'),
@@ -116,8 +117,13 @@ to_sql.relop_select_columns <- function (x,
   }
   using <- calc_using_relop_select_columns(x,
                                            using = using)
+  qlimit = limit
+  if(!getDBOption(db, "use_pass_limit", TRUE)) {
+    qlimit = NULL
+  }
   subsql_list <- to_sql(x$source[[1]],
                         db = db,
+                        limit = qlimit,
                         source_limit = source_limit,
                         indent_level = indent_level + 1,
                         tnum = tnum,
@@ -136,6 +142,10 @@ to_sql.relop_select_columns <- function (x,
          subsql, "\n",
          prefix, ") ",
          tab)
+  if(!is.null(limit)) {
+    q <- paste(q, "LIMIT",
+               format(ceiling(limit), scientific = FALSE))
+  }
   if(append_cr) {
     q <- paste0(q, "\n")
   }