+++ R/sql_node.R
@@ -197,6 +197,7 @@ prep_sql_toks <- function(db, ei) {
 to_sql.relop_sql <- function (x,
                               db,
                               ...,
+                              limit = NULL,
                               source_limit = NULL,
                               indent_level = 0,
                               tnum = mk_tmp_name_source('tsql'),
@@ -215,8 +216,13 @@ to_sql.relop_sql <- function (x,
   if(x$orig_columns) {
     cols <- c("*", cols)
   }
+  qlimit = limit
+  if(!getDBOption(db, "use_pass_limit", TRUE)) {
+    qlimit = NULL
+  }
   subsql_list <- to_sql(x$source[[1]],
                         db = db,
+                        limit = qlimit,
                         source_limit = source_limit,
                         indent_level = indent_level + 1,
                         tnum = tnum,
@@ -234,6 +240,10 @@ to_sql.relop_sql <- function (x,
               tab)
   if(!is.null(x$mods)) {
     q <- paste(q, x$mods)
+  }
+  if(!is.null(limit)) {
+    q <- paste(q, "LIMIT",
+               format(ceiling(limit), scientific = FALSE))
   }
   if(append_cr) {
     q <- paste0(q, "\n")