+++ NEWS.md
@@ -1,4 +1,11 @@
 
+# rquery 0.4.2 2018/04/05
+
+ * Work around RPostgreSQL issue (dbTableExists does not work).
+ * More controls on dbi_ steps through package options.
+ * Push LIMIT deeper into pipelines.
+ * Add connection config tools.
+
 # rquery 0.4.1 2018/03/17
 
  * Fix re-run/re-create situations.